Move user files
