+++ src_ts/components/app-shell/header/page-header.ts
@@ -10,7 +10,7 @@ import '@unicef-polymer/etools-profile-dropdown/etools-profile-dropdown';
 import '@unicef-polymer/etools-dropdown/etools-dropdown';
 import '@unicef-polymer/etools-app-selector/etools-app-selector';
 import '../header/countries-dropdown';
-import ProfileOperationsMixin from '../../user/profile-operations-mixin';
+import ProfileOperationsMixin from '../../common/user/profile-operations-mixin';
 import {isJsonStrMatch} from '../../utils/utils';
 import {fireEvent} from '../../utils/fire-custom-event';
 import {GenericObject, LabelAndValue, MinimalUser, User} from '@unicef-polymer/etools-types';
